Make layout preparations for dashboard styling
_app.js:
The firebase initialization was causing some errors. I commented this
out to keep working, but we'll have to deal with this eventually.

_document.js:
Added the style global tag here because it is generally true that we
want the body and the generated divs to take up the whole vertical space

There's still work to do with regards to the header growing in flex and
it is most likely that the main container will not be a flex, but rather
will host a Grid. Still unsure.