Add error handling to Admin page fetch
To our knowledge, the only possible error is a 500, which is produced
when the Team in question does not have an array for users. So we simply
return the empty array in the .then call, and handle it as if it were
empty.

On the other hand, if there is a network error, we present an unclosable
dialog that has a refresh button.